% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianWeather-data.R
\docType{data}
\name{CanadianWeather}
\alias{CanadianWeather}
\title{Canadian Weather dataset.}
\format{
A list that contains 8 variables:
\describe{
\item{dailyAvg}{average daily temperature and precipitation and log of precipitation for 35 Canadian weather stations}
\item{place}{places where Canadian weather stations are located}
\item{province}{provinces where Canadian weather stations are located}
\item{coordinates}{geographical coordinates}
\item{region}{regions where Canadian weather stations are located}
\item{monthlyTemp}{monthly temperature data}
\item{monthlyPrecip}{monthly precipitation data}
\item{geogindex}{geographic index}
}
}
\usage{
CanadianWeather
}
\description{
A dataset containing information about weather in Canadian weather stations
}
\keyword{datasets}
