% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Estimation of gradient of log-likelihood function}
\usage{
gradient(x, y, beta)
}
\arguments{
\item{x}{a design matrix which is a product of inner product of basis functions and basis coefficients of functional covariate \code{X}}

\item{y}{a response variable of class \code{factor}}

\item{beta}{initial values for beta regression coefficients and intercepts}
}
\value{
\item{grd}{  a vector of gradient values at the estimated optimum}
}
\description{
First derivative of log-likelihood function
}
