% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hessian.R
\name{Hessian}
\alias{Hessian}
\title{Estimation of Hessian matrix}
\usage{
Hessian(x, y, beta)
}
\arguments{
\item{x}{a design matrix which is a product of inner product of basis functions and basis coefficients of functional covariate \code{X}}

\item{y}{a response variable of class \code{factor}}

\item{beta}{initial values for beta regression coefficients and intercepts}
}
\value{
\item{Hess}{  a Hessian matrix at the estimated optimum}
}
\description{
Second derivative of log-likelihood function
}
