% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood function}
\usage{
loglik(x, y, beta)
}
\arguments{
\item{x}{a design matrix which is a product of inner product of basis functions and basis coefficients of functional covariate \code{X}}

\item{y}{a response variable of class \code{factor}}

\item{beta}{initial values for beta regression coefficients and intercepts}
}
\value{
\item{ll}{  a value of the log-likelihood function at the estimated optimum}
}
\description{
Log-likelihood function
}
