% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_freg.R
\name{summary.freg}
\alias{summary.freg}
\title{Summary of FREG model}
\usage{
\method{summary}{freg}(object, ...)
}
\arguments{
\item{object}{FREG model}

\item{...}{additional arguments relevant for the generic method}
}
\value{
call of the function

beta regression coefficient functions
}
\description{
\code{summary.freg} produce summary of FREG model fitting function \code{freg}.
}
