% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lfreg.R
\name{summary.lfreg}
\alias{summary.lfreg}
\title{Summary of LFREG model}
\usage{
\method{summary}{lfreg}(object, ...)
}
\arguments{
\item{object}{LFREG model}

\item{...}{additional arguments relevant for the generic method}
}
\value{
call of the function

beta regression coefficient functions
}
\description{
\code{summary.lfreg} produce summary of LFREG model fitting function \code{lfreg}.
}
