% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_olfreg.R
\name{summary.olfreg}
\alias{summary.olfreg}
\title{Summary of OLFREG model}
\usage{
\method{summary}{olfreg}(object, ...)
}
\arguments{
\item{object}{OLFREG model}

\item{...}{additional arguments relevant for the generic method}
}
\value{
call of the function

alpha regression coefficients and beta regression coefficient functions
}
\description{
\code{summary.olfreg} produce summary of OLFREG model fitting function \code{olfreg}.
}
